requirements.txt
django==4.1.3
pymysql
mysqlclient==2.1.0

django-cors-headers

pytest
pytest-django
coverage
pyyaml==6.0
djangorestframework~=3.13.1
requests

uwsgi

# 解密飞书事件
pycryptodome